worker/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
celery
redis
python-multipart
requests
python-dotenv
google-generativeai
moviepy
pillow
jsonschema
